Adding ngAnimate Support
Added support for ngAnimate by waiting for animations to complete
before focusing on the search input field. This uses $injector to get a
reference to $animate so that the project is not dependent upon
ngAnimate. Check plunkr http://plnkr.co/edit/G02EbD for a working
example.

This should resolve issues with ngAnimate 1.4.4 and 1.4.5. This should
cover the following issues:

https://github.com/angular-ui/ui-select/issues/859
https://github.com/angular-ui/ui-select/issues/1253