fix(parser) allow track by without filter
Previously the parser always attempted to extract a filter from the repeat
expression, regardless of whether it was present or not. This resulted in
the track by part of the expression being interpreted as a filter.

This commit changes the parser regex to capture the entire object source,
including its filters, before matching any track by expression. The filters
are then extracted separately from the source as this step is only required
when using an object as a source.

Fixes #1233